frontend: Make SearchResults a first-class component.
