+++ packages/frontend-web/src/app/scenes/Search/index.ts
@@ -1,5 +1,5 @@
 /*
-Copyright 2017 Google Inc.
+Copyright 2019 Google Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -19,110 +19,36 @@ import { withRouter } from 'react-router';
 import { compose } from 'redux';
 import { createStructuredSelector } from 'reselect';
 
-import { ICommentAction } from '../../../types';
 import { IAppDispatch, IAppStateRecord } from '../../stores';
 import { getArticleMap } from '../../stores/articles';
-import {
-  approveComments,
-  deferComments,
-  highlightComments,
-  rejectComments,
-  resetComments,
-  tagCommentSummaryScores,
-} from '../../stores/commentActions';
-import {
-  approveComment,
-  deferComment,
-  highlightComment,
-  rejectComment,
-  resetComment,
-} from '../../stores/comments';
-import { getTaggableTags } from '../../stores/tags';
-import { getTextSizes } from '../../stores/textSizes';
-import { Search as PureSearch } from './Search';
+import {ISearchProps, Search as PureSearch} from './Search';
 import {
   executeCommentListLoader,
   getAllCommentIds,
-  getAreAllSelected,
-  getAreAnyCommentsSelected,
-  getCommentListIsLoading,
-  getCurrentPagingIdentifier,
-  getIsItemChecked,
-  getSelectedCount,
   resetCommentIds,
-  toggleSelectAll,
-  toggleSingleItem,
 } from './store';
 import { ISearchScope } from './types';
 
 export { SearchResults } from './components/SearchResults';
 export { searchReducer } from './store';
 
-export interface IActionMap {
-  [key: string]: (ids: Array<string>, tagId?: string) => () => Promise<void>;
-}
-
-const updateCommentStateAction: {
-  [key: string]: any;
-} = {
-  highlight: highlightComment,
-  approve: approveComment,
-  defer: deferComment,
-  reject: rejectComment,
-  reset: resetComment,
-};
-
-const actionMap: IActionMap = {
-  highlight: highlightComments,
-  approve: approveComments,
-  defer: deferComments,
-  reject: rejectComments,
-  tag: tagCommentSummaryScores,
-  reset: resetComments,
-};
-
 const mapStateToProps = createStructuredSelector({
   totalCommentCount: (state: IAppStateRecord) => getAllCommentIds(state).size,
-  isLoading: (state: IAppStateRecord) => getCommentListIsLoading(state),
-  isItemChecked: (state: IAppStateRecord) => (id: string) => getIsItemChecked(state, id),
-  areNoneSelected: getAreAnyCommentsSelected,
-  areAllSelected: getAreAllSelected,
-  selectedCount: getSelectedCount,
   allCommentIds: getAllCommentIds,
-  tags: getTaggableTags,
   articleMap: getArticleMap,
-  textSizes: getTextSizes,
-  pagingIdentifier: getCurrentPagingIdentifier,
 });
 
-function mapDispatchToProps(dispatch: IAppDispatch): any {
+function mapDispatchToProps(dispatch: IAppDispatch): Partial<ISearchProps> {
   return {
     onSearch: async (newScope: ISearchScope)  => {
       dispatch(executeCommentListLoader(newScope));
     },
 
-    tagComments: (ids: Array<string>, tagId: string) =>
-        dispatch(tagCommentSummaryScores(ids, tagId)),
-
-    dispatchAction: (action: ICommentAction, idsToDispatch: Array<string>) =>
-        dispatch(actionMap[action](idsToDispatch)),
-
-    onToggleSelectAll: () => (
-      dispatch(toggleSelectAll())
-    ),
-
-    onToggleSingleItem: (item: { id: string }) => (
-      dispatch(toggleSingleItem(item))
-    ),
-
-    updateCommentState: (action: ICommentAction, ids: Array<string>) =>
-      dispatch(updateCommentStateAction[action](ids)),
-
     resetCommentIds: () => dispatch(resetCommentIds()),
   };
 }
 
-export const Search: React.ComponentClass = compose(
+export const Search: React.ComponentClass<{}> = compose(
   withRouter,
   connect(mapStateToProps, mapDispatchToProps),
 )(PureSearch);
